store_monitor_bot/requirements.txt
# ======================================================
# متطلبات المشروع (Requirements)
# ======================================================
# تثبيت كل شيء بأمر واحد:
#   pip install -r requirements.txt

# بوت تيليغرام
aiogram>=3.26.0

# قاعدة البيانات (PostgreSQL)
sqlalchemy==2.0.25
asyncpg==0.29.0          # driver غير متزامن لـ PostgreSQL
alembic==1.13.1           # للـ migrations (تغيير هيكل الجداول)

# السكرابينج
httpx==0.26.0             # طلبات HTTP غير متزامنة
beautifulsoup4==4.12.3    # تحليل HTML
lxml==5.1.0               # محلل HTML أسرع
playwright==1.41.1        # متصفح كامل للمواقع الصعبة
scrapling[fetchers]>=0.4.2  # HTTP + stealth/dynamic fetch (curl_cffi + patchright)

# Cache & Queue
redis==5.0.1              # للـ FSM storage وCache
celery==5.3.6             # لجدولة المهام في الإنتاج

# Admin Dashboard
fastapi==0.109.1
uvicorn[standard]==0.27.0
jinja2==3.1.3             # templates للداشبورد

# Security
python-jose[cryptography]==3.3.0  # JWT tokens
slowapi==0.1.9                    # rate limiting
bcrypt==4.1.3

# أدوات مساعدة
python-dotenv==1.0.0      # لقراءة ملف .env
pydantic>=2.12.0           # للتحقق من البيانات (updated for Python 3.14)
python-multipart==0.0.7   # لرفع الملفات في FastAPI

# اختياري - لتشغيل playwright
# playwright install chromium
